pythonDBconnect/requirements.txt
flask-sqlalchemy
pymysql
flask==2.2
